requirements.txt
google-cloud
surprise
storage